// apps/mobile/src/api/generated/communityOut.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hyperlocal Kitchen API
 * OpenAPI spec version: 0.1.0
 */
import type { CommunityType } from './communityType';
import type { CommunityZoneOut } from './communityZoneOut';

export interface CommunityOut {
  type: CommunityType;
  id: string;
  name: string;
  address: string | null;
  city: string;
  postal_code: string | null;
  status: string;
  zones: CommunityZoneOut[];
}